requirements.txt
Django==1.8.2
Jinja2==2.7.3
cached-property==1.2.0
celery==3.1.18
#celery-pipeline==0.15
djangorestframework==3.1.3
github3.py==1.0.0a2
jenkinsapi==0.2.28
#TODO: remove jsonfield
jsonfield==1.0.3
django-yamlfield==1.0.3
jsonpickle==0.9.2
redis==2.10.3
six==1.9.0
tox==2.1.1
munch
commandsession==0.3
wheel
